Move rmf data location
