+++ metrics-service/src/main/java/org/zowe/apiml/metrics/controllers/MetricServiceController.java
@@ -3,7 +3,7 @@ package org.zowe.apiml.metrics.controllers;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.zowe.apiml.metrics.RmfData;
+import org.zowe.apiml.metrics.services.zebra.RmfData;
 import org.zowe.apiml.metrics.services.zebra.ZebraMetricsService;
 
 import javax.ws.rs.QueryParam;
